# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/event/{id}": {
      "get": {
        "tags": [
          "event-controller"
        ],
        "operationId": "getEvent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EventDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "event-controller"
        ],
        "operationId": "updateEvent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "event-controller"
        ],
        "operationId": "deleteBuilding",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/building/{id}": {
      "get": {
        "tags": [
          "building-controller"
        ],
        "operationId": "getBuilding",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BuildingDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "building-controller"
        ],
        "operationId": "updateBuilding",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuildingDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "building-controller"
        ],
        "operationId": "deleteBuilding_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/event/": {
      "get": {
        "tags": [
          "event-controller"
        ],
        "operationId": "getEvents",
        "parameters": [
          {
            "name": "title",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "event-controller"
        ],
        "operationId": "createNewEvent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UserDto": {
        "required": [
          "firstName",
          "id",
          "lastName",
          "userRole",
          "username"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "userRole": {
            "type": "string",
            "enum": [
              "ADMIN",
              "ORGANIZER",
              "USER"
            ]
          },
          "username": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "pattern": "^\\+?[0-9. ()-]{7,25}$",
            "type": "string"
          },
          "age": {
            "maximum": 120,
            "minimum": 18,
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "BuildingDto": {
        "required": [
          "address"
        ],
        "type": "object",
        "properties": {
          "id": {
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "address": {
            "type": "string"
          },
          "hourlyRate": {
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "areaM2": {
            "minimum": 10,
            "type": "integer",
            "format": "int32"
          },
          "capacity": {
            "minimum": 10,
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "maxLength": 500,
            "minLength": 0,
            "pattern": "^[A-Za-zА-Яа-я0-9 ,.:;-]+|^$",
            "type": "string"
          },
          "rating": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BuildingRatingDto"
            }
          }
        }
      },
      "BuildingRatingDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "rating": {
            "type": "string",
            "format": "byte"
          },
          "author": {
            "$ref": "#/components/schemas/User"
          },
          "comment": {
            "type": "string"
          }
        }
      },
      "EventDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "eventTitle": {
            "type": "string"
          },
          "dateTimeStart": {
            "type": "string",
            "format": "date-time"
          },
          "dateTimeEnd": {
            "type": "string",
            "format": "date-time"
          },
          "building": {
            "$ref": "#/components/schemas/BuildingDto"
          },
          "description": {
            "type": "string"
          },
          "numberOfTickets": {
            "type": "integer",
            "format": "int32"
          },
          "minAgeRestriction": {
            "type": "integer",
            "format": "int32"
          },
          "rating": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventRatingDto"
            }
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserDto"
            }
          }
        }
      },
      "EventRatingDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "event": {
            "$ref": "#/components/schemas/EventDto"
          },
          "rating": {
            "type": "string",
            "format": "byte"
          },
          "author": {
            "$ref": "#/components/schemas/UserDto"
          },
          "comment": {
            "type": "string"
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "userRole": {
            "type": "string",
            "enum": [
              "ADMIN",
              "ORGANIZER",
              "USER"
            ]
          },
          "username": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "age": {
            "type": "integer",
            "format": "int32"
          }
        }
      }
    }
  }
}